triage_project/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.2
attrs==21.2.0
autopep8==1.5.6
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
decorator==4.4.2
Django==3.2
django-extensions==3.1.3
djangorestframework==3.12.4
drf-spectacular==0.20.1
Flask==1.1.2
idna==2.10
inflection==0.5.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==4.0.1
jupyter-client==6.1.11
jupyter-core==4.7.1
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.1
pathspec==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.16
pycodestyle==2.7.0
Pygments==2.8.0
pylint==2.7.4
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
